Handling repeat in run instead of cmd, env check only for spec level commands #862
