+++ cmd/docs.go
@@ -20,6 +20,7 @@ package cmd
 import (
 	"github.com/getgauge/gauge/api"
 	"github.com/getgauge/gauge/config"
+	"github.com/getgauge/gauge/env"
 	"github.com/getgauge/gauge/logger"
 	"github.com/getgauge/gauge/plugin"
 	"github.com/getgauge/gauge/track"
@@ -32,6 +33,9 @@ var docsCmd = &cobra.Command{
 	Long:    `Generate documentation using specified plugin.`,
 	Example: "  gauge docs spectacle specs/",
 	Run: func(cmd *cobra.Command, args []string) {
+		if e := env.LoadEnv(environment); e != nil {
+			logger.Fatalf(e.Error())
+		}
 		if err := config.SetProjectRoot(args); err != nil {
 			logger.Fatalf(err.Error())
 		}
